+++ deepcave/runs/converters/__init__.py
@@ -33,4 +33,6 @@ amltk
     This module provides utilities to create an AMLTK (AutoML Toolkit) run.
 optuna
     This module provides utilities to create an Optuna run.
+raytune
+    This module provides utilities to create a Raytune run.
 """