+++ CHANGELOG.md
@@ -1,3 +1,28 @@
+# Version 1.3.9
+
+## Parallel Coordinates
+- Set the margin of the graph so it gets dynamically computed depending on the length of the labels.
+
+# Version 1.3.6
+
+## New SMAC version compatibility
+- The smac3v2 converter can now also handle the output of the new smac version
+
+# Version 1.3.5
+
+## Converter
+- Added a new converter to handle RayTune runs
+- Added example RayTune runs to logs
+
+# Version 1.3.5
+
+## Access Specifier
+- Changed access specifier from '_fit' to 'fit' in the RF surrogate
+
+## Ablation tests
+- Added Polynomial "model" class for testing ablation path behavior
+- Added extra test case using the polynomial to validate ablation output correctness
+
 # Version 1.3.4
 
 ## Bug-Fixes